refactor(datafusion): move format table partition statements out of sql_context.rs
SHOW, ADD and DROP PARTITION for catalog-managed Format Tables now live in
format_partition_ddl.rs, the way DELETE, UPDATE and MERGE INTO have their
own modules. SQLContext keeps parsing and dispatch, and DROP PARTITION on a
Paimon table, which stays a snapshot commit.

The statement bodies and their helpers move unchanged. The SQLContext
helpers they share become pub(crate): resolve_catalog_and_table,
ensure_partition_command_target, partition_assignment, is_table_not_exist
and ok_result.
